requirements.txt
discord.py==2.2.2
python-dotenv==1.0.0
asgiref==3.6.0
openai==0.27.5
revChatGPT==5.0.0
GoogleBard==1.0.0
EdgeGPT==0.3.4.2
numpy==1.24.2
pandas==1.5.3
redis==4.5.1
requests==2.28.2
termcolor==2.2.0
jupyter==1.0.0
ipykernel==6.22.0
textract==1.6.5
tiktoken==0.3.3
streamlit==1.20.0